+++ src/CControl/Sources/Statistics/lda.c
@@ -18,13 +18,13 @@ static void center_data(float X[], float mu[], size_t row, size_t column);
  * W[m*c]
  * P[c*n]
  */
-void lda(const float X[], const size_t y[], float W[], float P[], const size_t c, const size_t row, const size_t column){
+void lda(const float X[], const uint8_t y[], float W[], float P[], const size_t c, const size_t row, const size_t column){
 	/* Create average vector mu_X = mean(X, 2) */
 	float* mu_X = (float*)malloc(row * sizeof(float));
 	average_vector(X, mu_X, row, column);
 
 	/* Count classes */
-	size_t amount_of_classes = y[column - 1] + 1;
+	uint8_t amount_of_classes = y[column - 1] + 1;
 
 	/* Create scatter matrices Sw and Sb */
 	size_t scatter_matrix_size = row * row;
